refactor: Memory optimizations, exception builder, application service
For unnecessary cases:
- change wrapper -> primitive
- remove toString()
Change exception responses from static constructor -> builder
Product application service has only one implementation. Thus, impl is removed and no interface is needed
